tirando parte lógica de orderModal e atribuindo a useOrderModal
